emacs-tree-sitter.org: use (require 'tree-sitter-langs)
The documentation as it is confusing: it basically says you have to
enable tree-sitter-mode, and stuff will just work. Well it actually
doesn't, because the variable `tree-sitter-major-mode-language-alist`
won't get set until you execute (require 'tree-sitter-langs).

And it is not clear at all that this line needs to be executed because
all you get is `No language registered for major mode` message.` Usually
lack of (require) results in missing function definitions, so in this
case there's no reason a user seeing the error would think "aha, I need
a (require) call".

So let's avoid confusion by mentioning the (require) line explicitly.