Add support for Android (with termux)
This change allows the package to work on Emacs on Android +
termux.

Users will still need to manually install the required termux
packages.

Additionally, the DYN-VERSION file in the `tsc` package (in
`.emacs.d/elpa/tsc-[version]/` ) may need to be manually modified to
conform to the expected format--e.g. from `emacs-tree-sitter@0.18.0`
to `0.18.0` (see
https://github.com/emacs-tree-sitter/elisp-tree-sitter/pull/249 for
more info).

Finally, a change to `tree-sitter-langs` package is also needed for
things to work locally (see
https://github.com/emacs-tree-sitter/tree-sitter-langs/pull/563).